Allow ActiveTool property to be undefined
No idea why this doesn't happen for the other properties in this QML element, but it seems to try to evaluate the 'visible' property upon closing the panel. If the panel is closed due to switching to a different tool (e.g. mirror tool) then the ActiveTool doesn't have this property any more and the evaluation fails.

Contributes to issue CURA-7501.