Only change pagesList selection after model is updated
There is an asynchronousity here that is kind of complex. We were inserting pages and immediately afterwards calling setPage(0) to select the first page. However inserting pages only inserts them into the configPagesModel. Inserting into the configPagesModel will eventually cause the pagesList to update through a signal. However it'll first complete the current setPage() function call before processing those signals. As a result, when we were changing the selection, the pagesList wasn't aware of any new pages yet.
To fix this, I'm making the selection change when the current page changes, and set the current page to 0 instead. This is then also a signal which gets processed after the signal that updates the pagesList. As a result, we don't get this QML error any more.

Contributes to issue CURA-7501.