Introduce the file provider plugin
The FileProvider plugin type can be used by plugins that want to insert an alternate way of
bringing files into Cura. They are handled by the FileProviderModel and once they are registered,
they will appear as options in the File->Open File(s) menu. The Open File(s) will be a simple
option in the Files menu when there are no other file providers. When there are multiple, then
the Open File(s) will be a sub-menu with all the file providers as options.

The "From Disk" option will always appear at the top of the file providers list.

CURA-7868