Revert making the OpenLocalFile dialog a separate LocalFileProvider plugin
Due to a bug in Qt, the FileDialog is not properly opened when it is triggered from a MenuItem
IF the FileDialog resides in its own qml file! Something is going wrong and the parent window is
not being fed as a context to the FileDialog window, hence the FileDialog never shows up.

This issue does not occur if the FileDialog exists as a component of the Cura main window.

According to our tests, this issue doesn't exist in Qt version >5.10. But since we use 5.10, we
decided to revert the changes that separate the OpenLocalFile dialog into a plugin. Once we upgrade
to a higher Qt version, we can revert this revert.

This reverts commits 8559f3f1733c10861e6cc898b7a69976f27962a6 to 591b67b7ccab4da735acffd4823e96c107490b88

CURA-7868