Catch FileNotFoundError when making temporary directory for version upgrade
If the user's hard disk is full this would crash. Or if he didn't have access to any of the temporary directories.
Instead we'll just not upgrade. If the hard disk was full, creating a new configuration will save some disk space (but probably crash later anyway).

Fixes Sentry issue CURA-1HT.