Add metadata to scene nodes
A spot where generic metadata can be stored. I considered making a ThreeMFMetadataDecorator class to store this in, but that would then be located in either the reader or the writer for 3MF, which is then inaccessible by the other one. I considered making a generic MetadataDecorator but then considered that scene node decorators by themselves are already kinda metadata. This would be overengineering. This is the simplest solution I can think of, and it won't be in the way of anything.

Contributes to issue CURA-7615.