Remove project output devices from the list when they get disabled
Now if a project output device gets disabled, it will be removed from the output_devices list if it
was necessary to be added there.

When it gets enabled again, the device is readded to this list, making it available to be selected
from the OutputDevicesActionButton in the bottom right.

CURA-7903