Fix slice info being send multiple times
The OutputDevice class is decorated with the signalemitter decorator, which allows it to declare
the instance signals outside the __init__ function. However, the ProjectOutputDevice was not
decorated similarly, which was causing it to have the writeSignals between ProjectOutputDevices
being shared. Therefore, when we emitted the writeStarted signal (which triggers the sending of
the SliceInfo) in the LocalFileOutputDevice, it was being triggered also for other
ProjectOutputDevices. Weird issue.

CURA-8164