Don't connect the signals if the device is disabled on startup
If the project output device was disabled at startup, the write signals were still connected. Then,
when the device was getting enabled, its write signals were getting connected again, leading to
issues like sending the slice info twice when the device was being used to write to a file.

This commit fixes that by making sure that the signals on startup are being connected ONLY if the
device is enabled.

CURA-8164