Use public functions of Resources class only
I considered exposing the data storage root path function publically but that's not workable since it only really works on Linux. You'd need to combine it with the config storage path... Well, it's just not designed to be exposed. It wouldn't be a nice public function.

Contributes to issue CURA-8162.