Allow for sub paths with getIcon
The new icons will be stored as multiple sizes. Best practice is to
store these in sub directories indicating the size, but keeping the
icon name the same.

A icon can now be called with `UM.Theme.getIcon('name', 'size')` the
default is 24 and that argument is therefore optional. For now it
has a fallback retrieve function that tries to grab the old icon from
the base directory.

CURA-8010_new_icons