Track MIME type of opened mesh files
This keeps track of each mesh what MIME type the mesh was loaded from, for the purpose of being able to send that information as anonymous usage statistic to Ultimaker.
It is accepted that meshes generated in other ways (e.g. from the Support Blocker plug-in) don't get an assigned MIME type.

Contributes to issue CURA-8232.