Add the total count of files in the directory during Trust check
When the central_storage.json contains a folder that is being moved to the central storage, then
Trust doesn't need to check for the signature of all the files in that folder. This is because
the `CentralFileStorage.retrieve()` already checks whether all those files have the correct hash.

In such cases, add all the files of that directory into the file_count.

CURA-8261