Process events while unzipping a plugin
Instead of calling the `extractall()` of the archive while unpacking a newly installed plugin, extract it file-by-file so that we can call the `processEvents()` after each extraction to make sure that the splash screen does not feel "stuck" if the extraction takes a long time.

This commit also adds a message to the splash screen to make it more verbose about which plugin it is currently unpacking.

CURA-8291