Restrict the locations of centrally stored files.
The central storage file shouldn't be able to point to files outside of its own folder. The original advice given here was to not use any (potential) directory delimeter characters, shuch as backslash or point(s), but that would a.) limit the use of the central file storage, since it then wouldn't be able to specify that sub-folders would need to be stored (or even files with extensions, if we just limited the use of '.' instead of '..') and b.) would be less robust if we forgot a character, like '~' can be used on some Linuxen to denote home-dir (mentioned in original advice, but just as an example). Instead, there is a method we already use in similar situations (which is a good reason in an of itself) to check this, so that is used instead. The method in question, PluginRegistry._isPathInLocation has been made more robust as well in this commit.

part of SEC-240 and its clone, CURA-8407