Central file storage + Trust: Add and fix unit-tests.
Added an extra unit-test for the new method (pre storage check). Also when the Central File Storage was originally written, and Turst adapted to be aware of that, the unit tests wherent changed to reflect the added functionality. Also, largely unrelated, but small enough, needed to change the mime-type test, as locally Qt isn't installed in English for some reason and then the 'description' is translated.

part of SEC-240 / CURA-8407