Refactor update checker to use the HttpRequestManager instead of requests
Although this wasn't technically a part of the ticket itself, it does
make the code a lot less complicated. This means that we no longer need
a seperate job for this to work. A further benefit is also that we
no longer use requests, which had some certifcate issues in the past.

CURA-7689