Make creation of shader program return None on failure.
Shaders are already treated (in frontend program) as if they where None on failure. Now actually make it so. This is part of a bugfix that came to our attention as 'If the slice button is clicked while the context-menu is also open on that same button, Cura crashes'. Suspect it may fix more than just that.

part of CURA-8412