Catch OSError
he store function can also throw OSError because it uses shutil.rmtree
and shutil.move and os.remove this isn't caught in here or in store we
check if the files exist, but it could also so it won't throw because
of that. But what about file, folders marked r/o or more likely
different owner with fewer rights. Can't hurt to catch this exception