Fix images not showing in messages on Windows
The image was not being displayed on Windows, even if you had an image_source. That was happening because the getImageSource was returning a string instead of a QUrl and that was not being converted properly in the MessageStack.

This commit fixes that by ensuring that the image source in the model will be a QUrl now.

CURA- 8497