Safely convert message image sources to QUrls
Some messages may already provide a QUrl as the image source. In such cases, Cura would crash because the `QUrl.fromLocalFile()` expects a string and not a QUrl as input. This is now accounted for to make sure that in the image source will not be converted to QUrl again.

CURA-8500