Remove containers.db from ignored files list
This file is stored in the cache now, so it won't get upgraded. Which is good, because we don't want to deal with a database with potentially outdated structure here. We'd have to write version upgrades for that database whenever we need to change the database structure, e.g. to add/remove/rename/recombine tables and fields.

Contributes to issue CURA-6069.