Catch ValueError and log errors occurring if config dir can't be made
ValueError can occur if there are weird characters in the path somehow.
We also want to see errors in the log if this directory fails to be created. However we automatically fail every single time because we create this every time we ask for a resource path, thousands of times in the start-up. So first check if it already exists before making it so that the happy path doesn't generate errors that it can't be created because it already exists.

Fixes Sentry issue CURA-2PR.