Add property to directly evaluate whether any packages are set for installing
This is used to show whether the user would need to restart Cura to finalise installation, update or removal of a package.
Instead of setting this property upon installing or uninstalling, we simply show it depending on whether there are actually any packages to update upon the next restart. Should be more reliable because it's not got any extra state.

Contributes to issue CURA-8587.