Prevent unneeded copy in container query
By adding the values of the medatada dict to a list, a complete copy of the list
is made. This isn't actually needed, since we just want to filter results (wich are then
copied to a list). As unneeded copies tend to be on the expensive side, this makes profile
loading a lot faster