Use Nonzero clipping, and only take 1st subpolygon
We use Nonzero just in case the user defined the polygon the wrong way around (e.g. in the setting for the print head shape in Cura).
We can take the 0th sub-polygon because the intersection of convex hulls should result in a single convex polygon too. It's impossible for that to result in multiple different polygons.

Contributes to issue CURA-8592.