Fix setting width and policy in weirder situations
If the parent is not a ScrollView, contentHeight doesn't exist so that is unreliable. Check for size instead.
Also, apparently setting the width is not reliable either. I'm not sure why but with the SettingView elements the width became 0 when it was undefined. I'm now affixing it to the theme entry. The padding there seems to be the default but I'll leave it like that.

Contributes to issue CURA-8686.