Add debug print statements around imports of Sip packages
These seem to crash on MacOS. I want to know which of them causes the crash. And to verify that it's indeed these. However at this stage we're not getting any log output yet because the logger hasn't been imported yet. We also don't get normal prints. Print to stderr, because that works.

Do not merge to master.