Workaround for bug in Windows version of Qt?
We want this to be both modal and a dialog/window with _only_ a close button, but we also want the (sub)window to be moveable w.r.t. Cura. The type being 'Qt.Dialog' would prevent the window from being movable on Linux. The type being Qt.Window (or even removing it) would prevent the close-button from showing up, _unless_ we also removed all hints. In which case the minimize and full-screen buttons would also be there, which is not ideal for a window that's supposed to be modal to say the least.

done as a 5-min fix as part of the whole Qt cleanup thing we're going through