Listen to return key to trigger default action in MessageDialog
As the key events didn't bubble through to the message dialog component we had to find a workaround; on opening of the `MessageDialog` set focus on the primary button. If a button has focus, then events _do_ get triggered. By triggering the click event on the return key pressed event.