Fix restoring window state upon restart
This fixes a TODO that was still in the code there.
In Qt6, the WindowState is a proper enum instead of an integer field in the Qt class. The setWindowState function expects that enum, not an integer. We were storing the integer in Cura 4, so that integer should be converted into the enum before calling setWindowState.

Additionally we were also converting the WindowState enum to a string when serialising it for the preferences class. This causes a problem with the version upgrade because the data in cura.cfg would need to change.

This change here fixes both. It makes the cura.cfg file store the integer as before so no version upgrade will be necessary for this. And it converts it properly to an enum for the setWindowState function.

Fixes issue CURA-9171.