Retrieved beta version would always be > local beta version.
Since we can't store the postfix (only up to 'revision' level) and major.minor.revision is loaded as 3 intergers from (online) latest.json -- the latest BETA version would always be (stoed online and) retrieved as a.b.c (withouth postfix info), but compared against the local a.b.c-beta+whatever (so, with postfix). The algorithm would then quite correctly deduce that a.b.c-beta+whatever was in fact a 'smaller' version than a.b.c (no postfix). Which then would lead to a message that a new beta was avialable _even if the user was already runnng said beta_.

This fixes that, but the underlying issue (not providing a postfix version) still remains. Problems this could cause are when we want to release and announce a second beta, without already incrementing the patch/revision nr. just because of that.

CURA-8856