Add option to get a short version of OpenGL version
Just the version number please. That's the one we'd like to know about: Which version of OpenGL the GPU pretends to support. That way we can do statistics on it.
The way it is now, we can't do statistics on it because the version number includes stuff like the build version of the firmware. That's not what we want to know. We'd like to know if the shaders fail to compile if the GPU doesn't support OpenGL4.1 or something.

Done as a 5 minute fix.