Correct fallback font weight to 500
In Qt6 these weights have been multiplied by 10. 500 is now a quite normal weight.
We don't have any fonts in our default theme that don't specify a weight, so it wasn't actually visible to the user. But if the theme is modified this should provide a better fallback.

Done as a 5-minute fix.