Don't allow containers without type and name in container registry
If the name was missing it would cause Cura to crash. If the type was missing it wouldn't crash, but it would silently fail to load the model which is also not good. With this change, in both cases it will add a warning to the log.

Contributes to issue CURA-3Z4.