Add a proper way for plugins to add settings.
Since they're loaded on top instead of going through the normal inheritance system, we need to merge/add them to each definition-container. This can't happen during load from file either, since most times, it's loaded from the cache-db instead. Instead, load the extra settings into the Definition ones each time one is loaded. If this turns out to be slow, we could pre-parse the json into (proto) SettingsDefinitions beforehand maybe, but it's not clear how much that is the actual bottleneck and how much code would need to be either duplicated or upended.

part of CURA-10722