Groups and merged meshes should also 'lay flat by face'.
Note that when this is the case (group or merged) the object selected for face node isn't the same as the 'overall' selected node in the scene. In that case, the overall object is the parent group or merged node, but the selected object 'for the face' needs to be one with a mesh directly attached to it (so a leaf in that tree).

part of CURA-10149