Rework object face selection
Instead of using a dynamic number of bits for the face index, store the object ID in the alpha channel, which was previously unused. This way we have the full 24 bits of the RGB channels to store the face index.
OpenGL blending is disabled for this pass, so the alpha channel is just stored as an extra channel but no calculation is performed based on it.

CURA-11984