For cross-stack relations, signal changed properties globally.
If a setting depends on a setting that is somehow in another stack (or another part of the global tree), as signalled by 'force_depends_on_settings' (renamed from 'depends_on_settings', as that'd imply you should _always_ add it), use the global container stack instead of the 'local' one (which'd only contain the setting-instances for that stack, not the actual relation _should_ it reside in another part of the currently active machine containers).

This can happen for instance in a dependee like Cura, which has machines with different extruders, and if a setting is locked to an extruder with 'limit_to_extruder', a setting using a formula using such a setting should be updated from what may potentially be the an other extruders' stack.

CURA-12050