Fix removing watching for file modifications
If the file is no longer part of the scene graph no longer watch it
for file modifications.  This had been triggered from MeshData when it
was deleted, but _application was no longer populated so it was not
being called.

_application could be replaced by querying the current instance.
from UM.Application import Application
Application.getInstance().getController().getScene().removeWatchedFile(self._file_name)

The problem with calling removeWatchedFile from MeshData __del__, is
Cura 5.8.1 is creating three MeshData objects with the file name
populated, when the file is loaded, two are deleted, the last stays
around.  If it were to call from the first two objects being deleted,
it would remove the watch even though it is loaded as part of the
scene.  Further I'm seeing that the last object isn't getting deleted,
even when it is removed from the scene graph.  I don't know if this is
a memory leak or if there would still be an issue with undo/redo.

Instead listen for the scene graph meshDataChanged signal, collect the
files in the scene graph and add or remove based on that list.  This
does not listen to sceneChanged signal as that includes transform
changes, which won't change the set of files listened for.