Update-checker: Fix display-same-version behaviour.
Previously, if there had been _any_ earlier succesful update-check, then subsequent versions would be ignored. That is (and this is maybe the reason we didn't catch this earlier?) unless the preferences weren't saved at some point that same session after the update-check was done -- in which case it'd always keep the older (already in the file) version in the preferences of the 'latest-update-shown' value. That's also fixed.

There's a unit-test that _almost_ would have caught this if they would've had enough entries in their 'try with these parameters' data. Also fixed now.

done as part of CURA-12650