Fix slice invalidation at file save
CURA-12652
When creating the merged instance contained at save time, we don't need/want property signals to be emitted. This is usually harmless, because the container is created locally and nobody is connected to it. However, when you have a setting with force_depends_on_settings, it will emit the signal using the real global stack instead, invalidating the slice.